+++ android/app/src/main/java/com/meter/giga/data/models/requests/TCPInfoRequestModel.kt
@@ -4,114 +4,114 @@ package com.meter.giga.data.models.requests
 import com.google.gson.annotations.SerializedName
 
 data class TCPInfoRequestModel(
-    @SerializedName("ATO")
-    val aTO: Int?,
-    @SerializedName("AdvMSS")
-    val advMSS: Int?,
-    @SerializedName("AppLimited")
-    val appLimited: Int?,
-    @SerializedName("Backoff")
-    val backoff: Int?,
-    @SerializedName("BusyTime")
-    val busyTime: Int?,
-    @SerializedName("BytesAcked")
-    val bytesAcked: Int?,
-    @SerializedName("BytesReceived")
-    val bytesReceived: Int?,
-    @SerializedName("BytesRetrans")
-    val bytesRetrans: Int?,
-    @SerializedName("BytesSent")
-    val bytesSent: Int?,
-    @SerializedName("CAState")
-    val cAState: Int?,
-    @SerializedName("DSackDups")
-    val dSackDups: Int?,
-    @SerializedName("DataSegsIn")
-    val dataSegsIn: Int?,
-    @SerializedName("DataSegsOut")
-    val dataSegsOut: Int?,
-    @SerializedName("Delivered")
-    val delivered: Int?,
-    @SerializedName("DeliveredCE")
-    val deliveredCE: Int?,
-    @SerializedName("DeliveryRate")
-    val deliveryRate: Int?,
-    @SerializedName("ElapsedTime")
-    val elapsedTime: Int?,
-    @SerializedName("Fackets")
-    val fackets: Int?,
-    @SerializedName("LastAckRecv")
-    val lastAckRecv: Int?,
-    @SerializedName("LastAckSent")
-    val lastAckSent: Int?,
-    @SerializedName("LastDataRecv")
-    val lastDataRecv: Int?,
-    @SerializedName("LastDataSent")
-    val lastDataSent: Int?,
-    @SerializedName("Lost")
-    val lost: Int?,
-    @SerializedName("MaxPacingRate")
-    val maxPacingRate: Int?,
-    @SerializedName("MinRTT")
-    val minRTT: Int?,
-    @SerializedName("NotsentBytes")
-    val notsentBytes: Int?,
-    @SerializedName("Options")
-    val options: Int?,
-    @SerializedName("PMTU")
-    val pMTU: Int?,
-    @SerializedName("PacingRate")
-    val pacingRate: Int?,
-    @SerializedName("Probes")
-    val probes: Int?,
-    @SerializedName("RTO")
-    val rTO: Int?,
-    @SerializedName("RTT")
-    val rTT: Int?,
-    @SerializedName("RTTVar")
-    val rTTVar: Int?,
-    @SerializedName("RWndLimited")
-    val rWndLimited: Int?,
-    @SerializedName("RcvMSS")
-    val rcvMSS: Int?,
-    @SerializedName("RcvOooPack")
-    val rcvOooPack: Int?,
-    @SerializedName("RcvRTT")
-    val rcvRTT: Int?,
-    @SerializedName("RcvSpace")
-    val rcvSpace: Int?,
-    @SerializedName("RcvSsThresh")
-    val rcvSsThresh: Int?,
-    @SerializedName("ReordSeen")
-    val reordSeen: Int?,
-    @SerializedName("Reordering")
-    val reordering: Int?,
-    @SerializedName("Retrans")
-    val retrans: Int?,
-    @SerializedName("Retransmits")
-    val retransmits: Int?,
-    @SerializedName("Sacked")
-    val sacked: Int?,
-    @SerializedName("SegsIn")
-    val segsIn: Int?,
-    @SerializedName("SegsOut")
-    val segsOut: Int?,
-    @SerializedName("SndBufLimited")
-    val sndBufLimited: Int?,
-    @SerializedName("SndCwnd")
-    val sndCwnd: Int?,
-    @SerializedName("SndMSS")
-    val sndMSS: Int?,
-    @SerializedName("SndSsThresh")
-    val sndSsThresh: Int?,
-    @SerializedName("SndWnd")
-    val sndWnd: Int?,
-    @SerializedName("State")
-    val state: Int?,
-    @SerializedName("TotalRetrans")
-    val totalRetrans: Int?,
-    @SerializedName("Unacked")
-    val unacked: Int?,
-    @SerializedName("WScale")
-    val wScale: Int?
+  @SerializedName("ATO")
+  val aTO: Long?,
+  @SerializedName("AdvMSS")
+  val advMSS: Long?,
+  @SerializedName("AppLimited")
+  val appLimited: Long?,
+  @SerializedName("Backoff")
+  val backoff: Long?,
+  @SerializedName("BusyTime")
+  val busyTime: Long?,
+  @SerializedName("BytesAcked")
+  val bytesAcked: Long?,
+  @SerializedName("BytesReceived")
+  val bytesReceived: Long?,
+  @SerializedName("BytesRetrans")
+  val bytesRetrans: Long?,
+  @SerializedName("BytesSent")
+  val bytesSent: Long?,
+  @SerializedName("CAState")
+  val cAState: Long?,
+  @SerializedName("DSackDups")
+  val dSackDups: Long?,
+  @SerializedName("DataSegsIn")
+  val dataSegsIn: Long?,
+  @SerializedName("DataSegsOut")
+  val dataSegsOut: Long?,
+  @SerializedName("Delivered")
+  val delivered: Long?,
+  @SerializedName("DeliveredCE")
+  val deliveredCE: Long?,
+  @SerializedName("DeliveryRate")
+  val deliveryRate: Long?,
+  @SerializedName("ElapsedTime")
+  val elapsedTime: Long?,
+  @SerializedName("Fackets")
+  val fackets: Long?,
+  @SerializedName("LastAckRecv")
+  val lastAckRecv: Long?,
+  @SerializedName("LastAckSent")
+  val lastAckSent: Long?,
+  @SerializedName("LastDataRecv")
+  val lastDataRecv: Long?,
+  @SerializedName("LastDataSent")
+  val lastDataSent: Long?,
+  @SerializedName("Lost")
+  val lost: Long?,
+  @SerializedName("MaxPacingRate")
+  val maxPacingRate: Long?,
+  @SerializedName("MinRTT")
+  val minRTT: Long?,
+  @SerializedName("NotsentBytes")
+  val notsentBytes: Long?,
+  @SerializedName("Options")
+  val options: Long?,
+  @SerializedName("PMTU")
+  val pMTU: Long?,
+  @SerializedName("PacingRate")
+  val pacingRate: Long?,
+  @SerializedName("Probes")
+  val probes: Long?,
+  @SerializedName("RTO")
+  val rTO: Long?,
+  @SerializedName("RTT")
+  val rTT: Long?,
+  @SerializedName("RTTVar")
+  val rTTVar: Long?,
+  @SerializedName("RWndLimited")
+  val rWndLimited: Long?,
+  @SerializedName("RcvMSS")
+  val rcvMSS: Long?,
+  @SerializedName("RcvOooPack")
+  val rcvOooPack: Long?,
+  @SerializedName("RcvRTT")
+  val rcvRTT: Long?,
+  @SerializedName("RcvSpace")
+  val rcvSpace: Long?,
+  @SerializedName("RcvSsThresh")
+  val rcvSsThresh: Long?,
+  @SerializedName("ReordSeen")
+  val reordSeen: Long?,
+  @SerializedName("Reordering")
+  val reordering: Long?,
+  @SerializedName("Retrans")
+  val retrans: Long?,
+  @SerializedName("Retransmits")
+  val retransmits: Long?,
+  @SerializedName("Sacked")
+  val sacked: Long?,
+  @SerializedName("SegsIn")
+  val segsIn: Long?,
+  @SerializedName("SegsOut")
+  val segsOut: Long?,
+  @SerializedName("SndBufLimited")
+  val sndBufLimited: Long?,
+  @SerializedName("SndCwnd")
+  val sndCwnd: Long?,
+  @SerializedName("SndMSS")
+  val sndMSS: Long?,
+  @SerializedName("SndSsThresh")
+  val sndSsThresh: Long?,
+  @SerializedName("SndWnd")
+  val sndWnd: Long?,
+  @SerializedName("State")
+  val state: Long?,
+  @SerializedName("TotalRetrans")
+  val totalRetrans: Long?,
+  @SerializedName("Unacked")
+  val unacked: Long?,
+  @SerializedName("WScale")
+  val wScale: Long?
 )
