+++ android/app/src/main/java/com/meter/giga/domain/entity/request/BBRInfoRequestEntity.kt
@@ -1,9 +1,9 @@
 package com.meter.giga.domain.entity.request
 
 data class BBRInfoRequestEntity(
-  val bw: Int?,
-  val cwndGain: Int?,
-  val elapsedTime: Int?,
-  val minRTT: Int?,
-  val pacingGain: Int?
+  val bw: Long?,
+  val cwndGain: Long?,
+  val elapsedTime: Long?,
+  val minRTT: Long?,
+  val pacingGain: Long?
 )
